+++ src/main/java/net/geforcemods/securitycraft/datagen/BlockTagGenerator.java
@@ -655,6 +655,7 @@ public class BlockTagGenerator extends BlockTagsProvider {
 				SCContent.RIFT_STABILIZER.get(),
 				SCContent.SCANNER_DOOR.get(),
 				SCContent.SCANNER_TRAPDOOR.get(),
+				SCContent.SECURE_TRADING_STATION.get(),
 				SCContent.SECURITY_CAMERA.get(),
 				SCContent.SONIC_SECURITY_SYSTEM.get(),
 				SCContent.TROPHY_SYSTEM.get(),